EUI-2852 - task field component (#696)
* Task field display

EUI-2852. The TaskFieldDisplay component, together with a test harness for QAs to assert that it behaves as described in the ticket.

There are currently no unit tests ready to commit for this but I didn't want this to get lost.

* Linting

EUI-2852. yarn lint threw up a few issues, which are now fixed. Also committing the yarn.lock file with the new version of rpx-xui-common-lib in place.

* Unit tests

EUI-2852. Bit of a refactor, following the introduction of a load of unit tests as the modules weren't behaving themselves the way I'd hoped. Also extracted out the URL and image components to make them independent and therefore more reusable.

Additionally, refactored the code around INTEGER and DECIMAL_2 types to elegantly handle non-numeric data.

* Linting

Linting caught me out again after some late changes.

* Build error

Weirdly, this wasn't showing locally but I can see the issue when Jenkins builds it - using an "object" in the test harness wasn't working so I've corrected that.

Also got rid of some lint warnings to tidy things up.

* Build error

Apparently, the imageRow now needs to be public.

Co-authored-by: Paul Graham <paul@smartcasual.io>