EUI-2735: My and Available tasks (PACT integration) (#772)
* EUI-2735

Refactored the PACT tests to be able to return individual tasks and lists of tasks, based on search criteria. The mechanism in place also supports "sorting" at the PACT level rather than having code in the UI to do it.

* Removed task field wrapper

Removed the now redundant task field wrapper component.

* PACT

Additional work on the PACT stubs to offer a series of tasks that will produce a predictable error when attempting to perform an action on them. There are also a series of caseworkers (in "Dodgy Location") that will similarly produce a predictable error when attempting to assign them to any of the tasks.

Should make it easier to test.

* Unit tests

Unit tests for new function in task field component.

Co-authored-by: Paul Graham <paul@smartcasual.io>