Improved handling of actions
The reassignment page now returns from whence it came when cancelling, succeeding, or with certain errors, rather than always going to My tasks.

Changed the mechanism for setting up actions on returned tasks so it's now done dynamically in the node layer, rather than being returned from the API itself.

This change to the mechanism also necessitated some changes in the way we make the calls so several Angular components and containers had to have their calls to the API modified.