Generic error handling
Added a utility function for handling fatal errors and diverting the user to an appopriate page where necessary. This function is now being called when searching tasks, loading locations, or loading caseworkers - and, at some point, should also be called when performing actions against tasks, but that work has been done in a separate branch so it can wait.