EUI-2913: Reassign task manager (#801)
* Refactor

Made the mechanism for getting back to the list from the reassignment page generic so it doesn't always return to My tasks.

* Improved handling of actions

The reassignment page now returns from whence it came when cancelling, succeeding, or with certain errors, rather than always going to My tasks.

Changed the mechanism for setting up actions on returned tasks so it's now done dynamically in the node layer, rather than being returned from the API itself.

This change to the mechanism also necessitated some changes in the way we make the calls so several Angular components and containers had to have their calls to the API modified.

* Update package.json

Shortcut to do run coverage analysis on both node and angular layers.

* If there is a 500, 401, or 403 error then we take the user to the relative error page.

* Merge

Additional adjustments and fixes necessary for the merge from workallocation-3.

* Error message

Added the error message component in from a now-redundant branch to finish off the unhappy path for reassignment.

* Update info-message-comms.service.ts

Lint warnings.

* Fixing up unit tests.

* Info messages in Task manager

Made an info message container with the logic that would otherwise have been in both the task home component and the task manager component and it's now available in both components abstracted.

* Fixing tests

The mock on one of the unit tests had become incomplete and resulted in failures.

* Update task-manager-list.component.ts

Don't remove the messages when selection changes in the Task manager as the selection changes when the URL includes the manage link hash, which means the message displays very briefly and then disappears.

* Generic error handling

Added a utility function for handling fatal errors and diverting the user to an appopriate page where necessary. This function is now being called when searching tasks, loading locations, or loading caseworkers - and, at some point, should also be called when performing actions against tasks, but that work has been done in a separate branch so it can wait.

* Task resolver errors

Handle fatal errors in the task resolver.

* Assign task page

Defaulting the assignment page to have a location the same as the one for the task being assigned and also set up the caseworker list to exclude the caseworker the current task is assigned to.

* Unhappy path tests added.

* Unhappy path tests added

* WA ngIntegration Unhappy path tests EUI-2913

* reorganised unhappy path test files

* ngIntegration tests for task assignment page unhappy path.

* Error handling

Work Allocation calls now handle 503 errors the same was as 500 errors. Also put handling in for failures to load locations and caseworkers in the task assignment component.

Updated the unit tests as appropriate.

* axios security

* removing the resolution of axios

* Other fatal errors

Redirect to the "Service down" page when failing to load work allocation filters and when failing to load the list of tasks.

* use chromedriver latest

* Update my-tasks.component.spec.ts

Test for default search request in My tasks now appears to be working consistently.

* Update task.resolver.ts

Failure to load a task should also be treated like a 500 error when clicking on, for example, "Reassign" in My tasks.

Co-authored-by: Paul Graham <paul@smartcasual.io>
Co-authored-by: Philip <philip.aberneithy@gmail.com>
Co-authored-by: sreekanthpuligadda <sreekanth.puligadda@hmcts.net>
Co-authored-by: Adnan Akgun <adnan.akgun@netclipper.co.uk>
Co-authored-by: Uday Denduluri <uday.denduluri@gmail.com>